+++ src/pudl/analysis/record_linkage/__init__.py
@@ -1,9 +1,8 @@
 """This module implements models for various forms of record linkage."""
 from . import (
     classify_plants_ferc1,
+    eia_ferc1_model_config,
     eia_ferc1_record_linkage,
-    eia_ferc1_record_linkage_splink_model,
-    eia_ferc1_splink_model_config,
     eia_ferc1_train,
     embed_dataframe,
     link_cross_year,
