+++ test/integration/ferc1_eia_train_test.py
@@ -19,7 +19,7 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from pudl.analysis.record_linkage.eia_ferc1_record_linkage import (
+from pudl.analysis.record_linkage.eia_ferc1_inputs import (
     restrict_train_connections_on_date_range,
 )
 from pudl.analysis.record_linkage.eia_ferc1_train import (
